move some runpaths tests to test_run_path_creation
